+++ stairway/game/button/FruitButton.scala
@@ -2,6 +2,7 @@
 package com.mercerenies.stairway.game.button
 
 import com.mercerenies.stairway.game.ButtonPad
+import com.mercerenies.stairway.util
 import com.mercerenies.stairway.util.Rectangle
 import com.mercerenies.stairway.action.KeyboardKey
 import java.awt.{Graphics2D, Font, Color}
@@ -13,6 +14,7 @@ class FruitButton(
   imageIndex: Int,
   hotkeys: Seq[KeyboardKey] = Nil)
     extends Button(pad, bbox, imageIndex, hotkeys) {
+  import util.GraphicsImplicits._
 
   private val font = new Font(Font.SANS_SERIF, Font.BOLD, 16)
   private var _count = initFruits
@@ -41,7 +43,7 @@ class FruitButton(
     val padding = 4
     val xPos = rect.xmax - padding - metrics.stringWidth(str)
     val yPos = rect.ymax - padding
-    graph.drawString(str, xPos.toInt, yPos.toInt)
+    graph.drawString(str, xPos, yPos)
   }
 
 }
